+++ src/main/kotlin/de/nordgedanken/auto_hotkey/lang/core/AhkFileType.kt
@@ -17,3 +17,5 @@ object AhkFileType : LanguageFileType(AhkLanguage) {
 
     override fun getDescription() = "${AhkConstants.LANGUAGE_NAME} script file"
 }
+
+fun VirtualFile.isAhkFile(): Boolean = fileType is AhkFileType