+++ src/main/kotlin/de/nordgedanken/auto_hotkey/runconfig/core/AhkRunConfig.kt
@@ -10,6 +10,7 @@ import com.intellij.execution.configurations.RuntimeConfigurationException
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.project.Project
+import de.nordgedanken.auto_hotkey.project.settings.defaultAhkSdk
 import de.nordgedanken.auto_hotkey.runconfig.execution.AhkRunState
 import de.nordgedanken.auto_hotkey.runconfig.model.AhkRunConfigSettings
 import de.nordgedanken.auto_hotkey.runconfig.ui.AhkRunConfigSettingsEditor
@@ -29,7 +30,12 @@ class AhkRunConfig(
 ) : LocatableConfigurationBase<RunProfileState>(project, factory, name) {
     var runConfigSettings = AhkRunConfigSettings()
 
-    override fun suggestedName(): String = runConfigSettings.pathToScript.substringAfterLast('/')
+    init {
+        // allows new configs to have the project ahk sdk set as the runner by default (will be overridden by template)
+        project.defaultAhkSdk?.let { runConfigSettings.runner = it.name }
+    }
+
+    override fun suggestedName() = "Run ${runConfigSettings.pathToScript.substringAfterLast('\\')}"
 
     override fun getConfigurationEditor(): SettingsEditor<out RunConfiguration?> {
         return AhkRunConfigSettingsEditor(this.project)
@@ -46,11 +52,11 @@ class AhkRunConfig(
     override fun checkConfiguration() {
         if (getAhkSdkByName(runConfigSettings.runner) == null) {
             throw RuntimeConfigurationError(AhkBundle.msg("runconfig.configtab.error.scriptrunner.notahksdktype"))
-        } else {
-            val f = File(runConfigSettings.pathToScript)
-            if (!f.exists()) {
+        }
+        File(runConfigSettings.pathToScript).run {
+            if (!exists()) {
                 throw RuntimeConfigurationError(AhkBundle.msg("runconfig.configtab.error.scriptpath.notexist"))
-            } else if (f.extension != AhkConstants.FILE_EXTENSION) {
+            } else if (extension != AhkConstants.FILE_EXTENSION) {
                 throw RuntimeConfigurationError(AhkBundle.msg("runconfig.configtab.error.scriptpath.notahkextension"))
             }
         }