+++ src/test/kotlin/de/nordgedanken/auto_hotkey/AhkProjectDescriptors.kt
@@ -1,12 +1,9 @@
 package de.nordgedanken.auto_hotkey
 
-import com.intellij.openapi.application.WriteAction
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.projectRoots.Sdk
-import com.intellij.openapi.projectRoots.SimpleJavaSdkType
 import com.intellij.openapi.projectRoots.impl.ProjectJdkImpl
-import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.testFramework.LightProjectDescriptor
+import de.nordgedanken.auto_hotkey.project.settings.defaultAhkSdk
 import de.nordgedanken.auto_hotkey.sdk.AhkSdkType
 
 /**
@@ -19,6 +16,7 @@ import de.nordgedanken.auto_hotkey.sdk.AhkSdkType
  */
 
 val mockAhkSdk = ProjectJdkImpl("Mock Ahk Sdk", AhkSdkType.getInstance())
+val mockAhkSdk2 = ProjectJdkImpl("Mock Ahk Sdk2", AhkSdkType.getInstance())
 
 object EmptyDescriptor : LightProjectDescriptor()
 
@@ -36,23 +34,8 @@ object WithOneAhkSdk : LightProjectDescriptor() {
 object WithOneAhkSdkAsProjDefault : LightProjectDescriptor() {
     override fun setUpProject(project: Project, handler: SetupHandler) {
         super.setUpProject(project, handler)
-        WriteAction.run<Exception> { ProjectRootManager.getInstance(project).projectSdk = mockAhkSdk }
+        project.defaultAhkSdk = mockAhkSdk
     }
 
     override fun getSdk() = mockAhkSdk
-}
-
-/**
- * ProjectDescriptor with a single java sdk added as the project's default
- */
-object WithOneJavaSdkAsProjDefault : LightProjectDescriptor() {
-    private lateinit var mockJavaSdk: Sdk
-
-    override fun setUpProject(project: Project, handler: SetupHandler) {
-        mockJavaSdk = ProjectJdkImpl("Mock Java Sdk", SimpleJavaSdkType.getInstance())
-        super.setUpProject(project, handler)
-        WriteAction.run<Exception> { ProjectRootManager.getInstance(project).projectSdk = mockJavaSdk }
-    }
-
-    override fun getSdk() = mockJavaSdk
 }